refactor: extract mutation definitions from resumes.ts into resumes_mutations.ts
Extract all write-operation mutations and internal queries from
resumes.ts (1171 → 625 lines, 47% reduction) into a new focused
resumes_mutations.ts module (592 lines).

Moved exports: updateAnalysis, updateAnalysisBatch,
updateIngestDataBatch, listResumeScanBatch, listResumeUsageBatch,
clearAnalyses, deleteResumes, archiveResumes, unarchiveResumes,
hardResetIngestData.

Also moved private helper normalizeRequestedResumeIds and type
definitions ResumeScanRow, ResumeUsageScanRow.

Updated all callers: BFF string-path references (resumes: →
resumes_mutations:), frontend api.resumes.* → api.resumes_mutations.*,
internal Convex references (internal.resumes.* →
internal.resumes_mutations.*).

Combined with prior extractions (diagnostics, search), resumes.ts is
now 625 lines — down from 2275 (69% total reduction).
