+++ apps/web/src/pages/DebugIngest.tsx
@@ -163,11 +163,11 @@ export default function DebugIngest() {
   const backfillIngestData = useAction(api.migrations.backfillIngestData)
   const reIngestStaleSkillsVersion = useAction(api.migrations.reIngestStaleSkillsVersion)
   const reIngestAllResumes = useAction(api.migrations.reIngestAllResumes)
-  const clearAnalysesMutation = useMutation(api.resumes.clearAnalyses)
-  const hardResetIngestDataMutation = useMutation(api.resumes.hardResetIngestData)
+  const clearAnalysesMutation = useMutation(api.resumes_mutations.clearAnalyses)
+  const hardResetIngestDataMutation = useMutation(api.resumes_mutations.hardResetIngestData)
   const resetDatabaseMutation = useMutation(api.resume_tasks.resetDatabase)
-  const deleteResumesMutation = useMutation(api.resumes.deleteResumes)
-  const archiveResumesMutation = useMutation(api.resumes.archiveResumes)
+  const deleteResumesMutation = useMutation(api.resumes_mutations.deleteResumes)
+  const archiveResumesMutation = useMutation(api.resumes_mutations.archiveResumes)
 
   const [search, setSearch] = useState('')
   const [skillsVersion, setSkillsVersion] = useState<number | null>(null)
